% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{calc_grad}
\alias{calc_grad}
\title{linear interpolation between a set of points; return gradient of each line}
\usage{
calc_grad(x, y)
}
\arguments{
\item{x}{numeric vector: x-coords of points.  Must be strictly increasing.}

\item{y}{numeric vector: y-coords of points.  Must be same length as x.}
}
\value{
numeric vector one shorter than x.
}
\description{
linear interpolation between a set of points; return gradient of each line
}
