% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{calc_LL_curve}
\alias{calc_LL_curve}
\title{summing log likelihood from saliva and each lymphocyte curve}
\usage{
calc_LL_curve(data_list, LL_row)
}
\arguments{
\item{data_list}{list with elements T_U (number of data points for saliva curve)
and T_L (vector of number of data points for each lymphocyte curve)}

\item{LL_row}{vector with elements named log_lik[i], each the log likelihood for a data point}
}
\value{
vector of length 1 + length(T_L): log likelihood for each curve
}
\description{
summing log likelihood from saliva and each lymphocyte curve
}
