% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{extract_pred}
\alias{extract_pred}
\title{extract model predictions as matrix}
\usage{
extract_pred(pred, N_samples, pred_name)
}
\arguments{
\item{pred}{stanfit object}

\item{N_samples}{number of samples used to make model predictions}

\item{pred_name}{(optional) name of model prediction.  If not specified, return everything}
}
\value{
matrix of model predictions
}
\description{
extract model predictions as matrix
}
