% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{filter_cell_data}
\alias{filter_cell_data}
\title{filter cell data frame based on conditions}
\usage{
filter_cell_data(cell_data, CD8_only = FALSE, NK_only = FALSE)
}
\arguments{
\item{cell_data}{data frame with cells column}

\item{CD8_only}{logical.  if TRUE, include CD8+ T cells only}

\item{NK_only}{logical. if TRUE, include NK cells only}
}
\value{
data frame
}
\description{
to do: be able to have more than one filtering condition be true
}
