% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{find_func_file}
\alias{find_func_file}
\title{find the file containing a function}
\usage{
find_func_file(func_name, dir_name = "~/git_repos/kirlabelling/R")
}
\arguments{
\item{func_name}{name of function}

\item{dir_name}{directory in which to search (only searches in that directory
+ subdirectories one level down)}
}
\value{
the name of the file which has the function
}
\description{
find the file containing a function
}
