% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_linear_stan.R
\name{fit_linear_stan}
\alias{fit_linear_stan}
\title{fit linear model to saliva, neutrophil and lymphocyte data for more than one participant}
\usage{
fit_linear_stan(
  ids,
  neutrophils,
  cells_lic.status,
  model,
  covariates,
  adapt_delta = 0.8,
  seed = 1,
  init = "random"
)
}
\arguments{
\item{ids}{participant ids}

\item{neutrophils}{neutrophil population (same for all ids)}

\item{cells_lic.status}{(optional) list of character vectors specifying the cells and
licensing statuses to include for each id.  Each element should be a subset of
get_id_data(id)$cells_lic.status}

\item{model}{either a stan model object used to fit the model to data,
or a string specifying model name to compile this model}

\item{covariates}{character vector including "iKIR_count", "lic.status", "cells",
"disease"; or a single value "all" or "null"}
}
\value{
list of two objects:
fit: stanfit object
data: data list used to fit stanfit object
}
\description{
fit linear model to saliva, neutrophil and lymphocyte data for more than one participant
}
