% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{fit_stan_defaults}
\alias{fit_stan_defaults}
\title{fit sta model with default parameters}
\usage{
fit_stan_defaults(
  model,
  data_list,
  init = "random",
  seed = 1,
  adapt_delta = 0.8,
  max_treedepth = 10,
  iter = 2000,
  evaluate_likelihood = TRUE
)
}
\arguments{
\item{model}{stan model object}

\item{data_list}{list of data to pass to model}

\item{init}{initialisation}

\item{adapt_delta}{scalar.  value of adapt_delta to use}

\item{evaluate_likelihood}{logical.  if TRUE, sample from posterior, if FALSE, sample from prior}
}
\value{
list of two objects:
fit: stanfit object
data: data list used to fit stanfit object
}
\description{
fit sta model with default parameters
}
