% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{get_cells_lic.status}
\alias{get_cells_lic.status}
\title{wrapper to only get cells_lic.status column for id data}
\usage{
get_cells_lic.status(id, CD8_only = FALSE, NK_only = FALSE)
}
\arguments{
\item{id}{string: participant id}

\item{CD8_only}{logical.  if TRUE, include CD8+ T cells only}

\item{NK_only}{logical. if TRUE, include NK cells only}
}
\value{
character vector
}
\description{
wrapper to only get cells_lic.status column for id data
}
