% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{get_fixed_pars}
\alias{get_fixed_pars}
\title{specify values of fixed parameters}
\usage{
get_fixed_pars(cells)
}
\arguments{
\item{cells}{(optional) if "Granulocytes" or "Monocytes" specified, return R and deltaT for these
cell populations}
}
\value{
named list of fixed parameters
}
\description{
specify values of fixed parameters
}
