% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{get_id_data}
\alias{get_id_data}
\title{get cell populations and licensing statuses for a given id in a lymphocyte study}
\usage{
get_id_data(id)
}
\arguments{
\item{id}{string: participant id}
}
\value{
data frame with five columns: cells, lic.status, cells_lic.status which
is the first two pasted together, disease ("HCV", "HIV-1", "HTLV-1" or "Control"), 
functional_iKIR_count(number of unique license genes)
}
\description{
get cell populations and licensing statuses for a given id in a lymphocyte study
}
