% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{get_ids}
\alias{get_ids}
\title{get ids in lymphocyte study}
\usage{
get_ids()
}
\arguments{
\item{threeDL2_neg}{logical. if TRUE, only get ids where 3DL2 is unlicensed}
}
\value{
character vector of ids
}
\description{
get ids in lymphocyte study
}
