% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{get_input_dir}
\alias{get_input_dir}
\title{set directory for input files}
\usage{
get_input_dir()
}
\value{
string specifying absolute path to directory containing input files
}
\description{
set directory for input files
}
