% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{get_max_LL_row}
\alias{get_max_LL_row}
\title{get the row with maximum likelihood from a stanfit object or}
\usage{
get_max_LL_row(fit_obj)
}
\arguments{
\item{fit_obj}{a stanfit object or matrix containing draws from posterior}
}
\value{
named vector with parameters and generated quantities of maximum likelihood
}
\description{
get the row with maximum likelihood from a stanfit object or
}
