% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{get_model_filename}
\alias{get_model_filename}
\title{get model filename}
\usage{
get_model_filename(model_name)
}
\arguments{
\item{model_name}{string}
}
\value{
string: model filename
}
\description{
get model filename
}
