% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{get_par_names}
\alias{get_par_names}
\title{get vector of parameter names added by each layer of the model}
\usage{
get_par_names()
}
\value{
named list of character vectors
}
\description{
get vector of parameter names added by each layer of the model
}
