% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{index_par_name}
\alias{index_par_name}
\title{append indices to parameter name}
\usage{
index_par_name(par_name, idx)
}
\arguments{
\item{par_name}{string: parameter name}

\item{idx}{numeric vector}
}
\value{
string.  e.g. if par_name == "abc" and idx == c(1,2), returns "abc[1,2]"
}
\description{
append indices to parameter name
}
