% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{interpolate_vector}
\alias{interpolate_vector}
\title{interpolate ordered numeric vector into n_div intervals betwen successive values}
\usage{
interpolate_vector(vec, n_div)
}
\arguments{
\item{vec}{numeric vector to interpolate}

\item{n_div}{numeric vector of length 1: number of intervals into which to subdivide
successive values of the vector}
}
\value{
numeric vector of length (length(vec) - 1) * n_div + 1: interpolated vector
}
\description{
interpolate ordered numeric vector into n_div intervals betwen successive values
}
