% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{make_filename}
\alias{make_filename}
\title{concatenate strings for filename}
\usage{
make_filename(strs, ext1 = "rds")
}
\arguments{
\item{strs}{list of characters}

\item{ext1}{string: extension}
}
\value{
string
}
\description{
concatenate strings for filename
}
