% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{make_pred}
\alias{make_pred}
\title{make model predictions}
\usage{
make_pred(model, data_list)
}
\arguments{
\item{model}{stan model object}

\item{data_list}{list of data and model parameters to feed to model}
}
\value{
stanfit object
}
\description{
make model predictions
}
