% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_frac_data}
\alias{read_frac_data}
\title{read data on the fraction of each cell population in each individual}
\usage{
read_frac_data(ids)
}
\arguments{
\item{ids}{character vector. participant ids}
}
\value{
tibble with columns id, cells, lic.status, frac.cells
}
\description{
read data on the fraction of each cell population in each individual
}
