% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_gm_data}
\alias{read_gm_data}
\title{read in granulocyte or monocyte data}
\usage{
read_gm_data(ids, cells)
}
\arguments{
\item{ids}{character vector. partiipant ids}

\item{cells}{"Granulocytes", "Monocytes" or "all".  Read data for these cell types}
}
\value{
data frame with four columns: id, time, Lb, cells
}
\description{
read in granulocyte or monocyte data
}
