% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_kir_data}
\alias{read_kir_data}
\title{read in kir data}
\usage{
read_kir_data(ids)
}
\arguments{
\item{ids}{character vector. participant ids}
}
\value{
data frame if length(ids) > 1, named vector otherwise
}
\description{
read in kir data
}
