% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_lymphocyte_data}
\alias{read_lymphocyte_data}
\title{read in lymphocyte data}
\usage{
read_lymphocyte_data(ids, cells_lic.status, drop_cells = FALSE)
}
\arguments{
\item{ids}{character vector. participant ids}

\item{drop_cells}{logical.  if TRUE, output three columns: id, time, L.
if FALSE, output five columns: id, time, L, cells, lic.status}

\item{cells_lic_status}{(optional) either data frame with two columns/list wih two elements: cells and lic.status,
or vector which is paste(cells, lic.status)
If specified, read those cells and license statuses only.  If not specified, read all.}
}
\value{
data frame
}
\description{
read in lymphocyte data
}
