% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_saliva_data}
\alias{read_saliva_data}
\title{read in saliva data}
\usage{
read_saliva_data(ids)
}
\arguments{
\item{ids}{(optional) character vector. if given, subset data for these ids only}
}
\value{
data frame of saliva data
}
\description{
read in saliva data
}
