% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{rep_each}
\alias{rep_each}
\title{repeat elements of vector x, y times each element-wise, where y is a vector}
\usage{
rep_each(x, y)
}
\arguments{
\item{x}{vector}

\item{y}{vector of same length as x}
}
\value{
vector of length sum(y)
}
\description{
repeat elements of vector x, y times each element-wise, where y is a vector
}
