% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{run_for_id}
\alias{run_for_id}
\title{evaluate a function for all cell populations and licensing statuses for a given id}
\usage{
run_for_id(f, id1, ...)
}
\arguments{
\item{f}{function to run.  First two arguments are cells and lic.status}

\item{id1}{id to run for}

\item{...}{other arguments to pass to f}
}
\value{
list of outputs for all cell populations and licensing statuses
}
\description{
evaluate a function for all cell populations and licensing statuses for a given id
}
