% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{run_ids_label}
\alias{run_ids_label}
\title{evaluate a function for all cell populations and licensing statuses for given ids,
and label outputs with id, cell population and licensing status}
\usage{
run_ids_label(f, ids)
}
\arguments{
\item{f}{function to run.  First two arguments are cells and lic.status}

\item{ids}{ids to run for}

\item{...}{other arguments to pass to f}
}
\value{
data frame of outputs for ids, cell populations and licensing statuses
}
\description{
evaluate a function for all cell populations and licensing statuses for given ids,
and label outputs with id, cell population and licensing status
}
