% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{subchunkify}
\alias{subchunkify}
\title{Change fig.width and fig.height Dynamically Within an R Markdown Chunk
http://michaeljw.com/blog/post/subchunkify/}
\usage{
subchunkify(g, fig_height = 7, fig_width = 5)
}
\arguments{
\item{g}{a ggplot object}

\item{fig_height}{scalar: figure height}

\item{fig_width}{scalar: figure width}
}
\value{
gpglot object with correct height and width
}
\description{
Change fig.width and fig.height Dynamically Within an R Markdown Chunk
http://michaeljw.com/blog/post/subchunkify/
}
