% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{subsample}
\alias{subsample}
\title{subsample elements of vector with constant spacing}
\usage{
subsample(vec, length_out)
}
\arguments{
\item{vec}{vector}

\item{length_out}{number of samples}
}
\value{
vector of length length_out
}
\description{
subsample elements of vector with constant spacing
}
