% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{vcapply}
\alias{vcapply}
\title{short for \code{vapply(X, FUN, character(1), ...)}}
\usage{
vcapply(X, FUN, ...)
}
\arguments{
\item{X}{first input to vapply}

\item{FUN}{second input to vapply}
}
\value{
output of \code{vapply(X, FUN, character(1), ...)}: character vector
}
\description{
short for \code{vapply(X, FUN, character(1), ...)}
}
