% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{vectorise_lymphocyte_pars}
\alias{vectorise_lymphocyte_pars}
\title{wrapper to vectorise lymphocyte parameter names}
\usage{
vectorise_lymphocyte_pars(
  pop,
  scalar_pars = c("frac", "delta"),
  vector_pars = c("pb_w", "dstar", "delay")
)
}
\arguments{
\item{pop}{number of lymphocyte populations}

\item{scalar_pars}{character vector.  names of scalar parameters}

\item{vector_pars}{character vector.  names of vector parameters}
}
\value{
names of scarla and vectorised parameters
}
\description{
wrapper to vectorise lymphocyte parameter names
}
