% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{viapply}
\alias{viapply}
\title{short for \code{vapply(X, FUN, integer(1), ...)}}
\usage{
viapply(X, FUN, ...)
}
\arguments{
\item{X}{first input to vapply}

\item{FUN}{second input to vapply}
}
\value{
output of \code{vapply(X, FUN, integer(1), ...)}: integer vector of 
same length as X
}
\description{
short for \code{vapply(X, FUN, integer(1), ...)}
}
