% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{vlapply}
\alias{vlapply}
\title{short for \code{vapply(X, FUN, logical(1), ...)}}
\usage{
vlapply(X, FUN, ...)
}
\arguments{
\item{X}{first input to vapply}

\item{FUN}{second input to vapply}
}
\value{
output of \code{vapply(X, FUN, logical(1), ...)}: logical vector of 
same length as X
}
\description{
short for \code{vapply(X, FUN, logical(1), ...)}
}
