% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{vnapply}
\alias{vnapply}
\title{short for \code{vapply(X, FUN, numeric(1), ...)}}
\usage{
vnapply(X, FUN, ...)
}
\arguments{
\item{X}{first input to vapply}

\item{FUN}{second input to vapply}
}
\value{
output of \code{vapply(X, FUN, numeric(1), ...)}: numeric vector of 
same length as X
}
\description{
short for \code{vapply(X, FUN, numeric(1), ...)}
}
